requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements/base.in
#
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via jsonschema
backports-zoneinfo==0.2.1
    # via django
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.1
    # via nbconvert
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   nltk
    #   unstructured-api-tools
    #   uvicorn
defusedxml==0.7.1
    # via nbconvert
deprecated==1.2.13
    # via limits
django==4.1.4
    # via image
entrypoints==0.4
    # via jupyter-client
fastapi==0.88.0
    # via
    #   -r requirements/base.in
    #   unstructured-api-tools
fastjsonschema==2.16.2
    # via nbformat
filelock==3.8.2
    # via
    #   huggingface-hub
    #   transformers
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.2
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.23.1
    # via -r requirements/base.in
huggingface-hub==0.11.1
    # via transformers
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
image==1.5.33
    # via -r requirements/base.in
importlib-metadata==5.1.0
    # via nbconvert
importlib-resources==5.10.1
    # via jsonschema
jinja2==3.1.2
    # via
    #   nbconvert
    #   unstructured-api-tools
joblib==1.1.0
    # via nltk
jsonschema==4.17.3
    # via nbformat
jupyter-client==7.4.8
    # via nbclient
jupyter-core==5.1.0
    # via
    #   -r requirements/base.in
    #   jupyter-client
    #   nbclient
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.2.2
    # via nbconvert
limits==2.7.2
    # via slowapi
lxml==4.9.1
    # via unstructured
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
mistune==2.0.4
    # via nbconvert
mypy==0.991
    # via unstructured-api-tools
mypy-extensions==0.4.3
    # via mypy
nbclient==0.7.2
    # via nbconvert
nbconvert==7.2.6
    # via unstructured-api-tools
nbformat==5.7.0
    # via
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.6
    # via jupyter-client
nltk==3.7
    # via unstructured
numpy==1.23.5
    # via transformers
packaging==22.0
    # via
    #   huggingface-hub
    #   limits
    #   nbconvert
    #   transformers
pandocfilters==1.5.0
    # via nbconvert
pillow==9.3.0
    # via image
pkgutil-resolve-name==1.3.10
    # via jsonschema
platformdirs==2.6.0
    # via jupyter-core
protobuf==3.20.3
    # via -r requirements/base.in
pydantic==1.10.2
    # via fastapi
pygments==2.13.0
    # via nbconvert
pyrsistent==0.19.2
    # via jsonschema
python-dateutil==2.8.2
    # via jupyter-client
python-dotenv==0.21.0
    # via uvicorn
python-multipart==0.0.5
    # via unstructured-api-tools
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
pyzmq==24.0.1
    # via jupyter-client
ratelimit==2.2.1
    # via -r requirements/base.in
regex==2022.7.9
    # via
    #   nltk
    #   transformers
requests==2.28.1
    # via
    #   -r requirements/base.in
    #   huggingface-hub
    #   transformers
rfc3986[idna2008]==1.5.0
    # via httpx
sentencepiece==0.1.97
    # via -r requirements/base.in
six==1.16.0
    # via
    #   bleach
    #   image
    #   python-dateutil
    #   python-multipart
slowapi==0.1.7
    # via unstructured-api-tools
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
starlette==0.22.0
    # via fastapi
tinycss2==1.2.1
    # via nbconvert
tokenizers==0.13.2
    # via transformers
tomli==2.0.1
    # via mypy
torch==1.13.0
    # via -r requirements/base.in
tornado==6.2
    # via jupyter-client
tqdm==4.64.0
    # via
    #   huggingface-hub
    #   nltk
    #   transformers
traitlets==5.7.0
    # via
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
transformers==4.25.1
    # via -r requirements/base.in
types-requests==2.28.11.5
    # via unstructured-api-tools
types-ujson==5.6.0.0
    # via unstructured-api-tools
types-urllib3==1.26.25.4
    # via types-requests
typing-extensions==4.4.0
    # via
    #   huggingface-hub
    #   limits
    #   mypy
    #   pydantic
    #   starlette
    #   torch
unstructured==0.2.5
    # via -r requirements/base.in
unstructured-api-tools==0.4.6
    # via -r requirements/base.in
urllib3==1.26.13
    # via requests
uvicorn[standard]==0.20.0
    # via unstructured-api-tools
uvloop==0.17.0
    # via uvicorn
watchfiles==0.18.1
    # via uvicorn
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websockets==10.4
    # via uvicorn
wrapt==1.14.1
    # via deprecated
zipp==3.11.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
